+++ dashboard/constants.py
@@ -72,3 +72,5 @@ CALENDAR_VERBS = ('BEGIN:VEVENT', 'END:VEVENT', 'BEGIN:VTODO', 'END:VTODO')
 GIT_PLATFORMS = ('GitHub', 'GitLab', 'Pagure')
 GIT_REPO_TYPE = ('default', 'localization', TRANSPLATFORM_ENGINES[3], TRANSPLATFORM_ENGINES[1],
                  TRANSPLATFORM_ENGINES[2])
+
+SYS_EMAIL_ADDR = "system@transtats.org"