+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 
 ##### Mon Jun 6 2022 Sundeep Anand <suanand@redhat.com>
+* Create default Pipeline Configurations
+* Redesign Package specific Add CI Pipeline form
 * Add target_langs and UniqueConstraint to pipeline_config model
 * Better pipeline history UI: group_by date order_by time dsc
 * Better error messages for exceptions in jobs
 * Executes pipeline configurations in threads
-* Redesign Package specific Add CI Pipeline form
 * fix pipeline push/pull actions for satellite
 * change repo_branch ui field to checkbox for satellite in pipeline config
 * fix issues in Add Release form